POST request & UI edge cases fixed + restController & form edit
