+++ frontend/src/App.js
@@ -40,8 +40,7 @@ import PrintBarcode from "./components/printBarcode/Index";
 import NonConformIndex from "./components/nonconform/index";
 import SampleBatchEntrySetup from "./components/batchOrderEntry/SampleBatchEntrySetup.js";
 import AuditTrailReportIndex from "./components/reports/auditTrailReport/Index.js";
-import AddOrganization from "./components/admin/OrganizationManagament/AddOrganization.js";
-import ModifyOrganization from "./components/admin/OrganizationManagament/ModifyOrganization.js";
+import OrganizationAddEdit from "./components/admin/OrganizationManagament/OrganizationAddModify.js";
 
 export default function App() {
   let i18nConfig = {
@@ -223,21 +222,15 @@ export default function App() {
                   role="Global Administrator"
                 />
                 <SecureRoute
+                  path="/Organization"
+                  exact
+                  component={() => <OrganizationAddEdit />}
+                  role="Global Administrator"
+                />
+                <SecureRoute
                   path="/MasterListsPage"
                   exact
                   component={() => <Admin />}
-                  role="Global Administrator"
-                />
-                <SecureRoute
-                  path="/AddOrganization"
-                  exact
-                  component={() => <AddOrganization />}
-                  role="Global Administrator"
-                />
-                <SecureRoute
-                  path="/ModifyOrganization"
-                  exact
-                  component={() => <ModifyOrganization />}
                   role="Global Administrator"
                 />
                 <SecureRoute
